+++ frontend/src/app/dashboard/dashboard.component.html
@@ -31,5 +31,5 @@
       </mat-select>
     </mat-form-field>
   </div>
-  <app-team-detail></app-team-detail>
+  <app-team-detail [teams]="teamList"></app-team-detail>
 </div>
