+++ frontend/src/app/team-detail/team-detail.component.html
@@ -1,6 +1,11 @@
-<div class="teamDiv" *ngFor="let teamObjective of teamObjectives">
-  <h1>{{ teamObjective.name }} Ziele</h1>
-  <div *ngFor="let objective of teamObjective.objectives">
+<div class="teamDiv" *ngFor="let team of teams | async">
+  <h1>{{ team.name }} Ziele</h1>
+  <div
+    *ngFor="
+      let objective of this.objectiveService.getObjectivesOfTeam(team.id)
+        | async
+    "
+  >
     <hr />
     <app-objective-row [objective]="objective"></app-objective-row>
   </div>
