+++ frontend/src/app/team-detail/team-detail.component.ts
@@ -1,23 +1,7 @@
-import { Component, OnInit } from '@angular/core';
-
-export interface Objective {
-  id: number;
-  title: String;
-  ownerId: number;
-  ownerFirstname: String;
-  ownerLastname: String;
-  quarterId: number;
-  quarterNumber: number;
-  quarterYear: number;
-  description: String;
-  progress: number;
-}
-
-export interface TeamObjective {
-  id: number;
-  name: string;
-  objectives: Objective[];
-}
+import { Component, Input, OnInit } from '@angular/core';
+import { Team } from '../dashboard/team.service';
+import { Observable } from 'rxjs';
+import { ObjectiveService } from './objective.service';
 
 @Component({
   selector: 'app-team-detail',
@@ -25,48 +9,9 @@ export interface TeamObjective {
   styleUrls: ['./team-detail.component.scss'],
 })
 export class TeamDetailComponent implements OnInit {
-  //Daten manuell einfügen
-  objectives: Array<Objective> = [
-    {
-      id: 3,
-      title: 'Wir erzielen eine ausgezeichnete Wirtschaftlichkeit',
-      ownerId: 2,
-      ownerFirstname: 'Rudi',
-      ownerLastname: 'Grochde',
-      quarterId: 5,
-      quarterNumber: 3,
-      quarterYear: 2022,
-      description: 'Sehr wichtig',
-      progress: 5,
-    },
-    {
-      id: 4,
-      title: 'Wir erreichen unsere anvisierten Wachstumsziele',
-      ownerId: 2,
-      ownerFirstname: 'Rudi',
-      ownerLastname: 'Grochde',
-      quarterId: 5,
-      quarterNumber: 3,
-      quarterYear: 2022,
-      description: 'Überaus wichtig',
-      progress: 30,
-    },
-  ];
-  teamObjectiveRuby: TeamObjective = {
-    id: 3,
-    name: 'dev/ruby',
-    objectives: this.objectives,
-  };
-  teamObjectiveBBT: TeamObjective = {
-    id: 4,
-    name: 'bbt',
-    objectives: this.objectives,
-  };
-  teamObjectives: TeamObjective[] = [
-    this.teamObjectiveRuby,
-    this.teamObjectiveBBT,
-  ];
-  constructor() {}
+  @Input()
+  teams!: Observable<Team[]>;
+  constructor(public objectiveService: ObjectiveService) {}
 
   ngOnInit(): void {}
 }
